feat(template-frontend): fix errors when creating rectangle on mouse moving

+++ src-tauri/src/watcher/window_query.rs
@@ -65,7 +65,7 @@ fn get_window_information_by_id(window_id: i64) -> Result<WindowInformation> {
         .duration_since(SystemTime::UNIX_EPOCH)
         .unwrap();
     let mut window_info = WindowInformation {
-        time: unix_ts.as_secs(),
+        time: unix_ts.as_secs() as u128,
         title: None,
         class: None,
         exec_path: None,
